requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=0.24.0
omegaconf>=2.1.0
wandb>=0.12.0
biopython>=1.79
fair-esm>=2.0.0
tqdm>=4.62.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
h5py>=3.0.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0
flake8>=3.9.0
isort>=5.9.0
jupyter>=1.0.0
tensorboard>=2.7.0
# Updated: 05/30/2025 22:59:09

# Updated: 05/31/2025 15:11:07

# Updated: 05/31/2025 15:14:04

# Updated: 05/31/2025 23:30:00

# Updated: 05/31/2025 23:30:18
